packages/pipeline/pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Evergreen M365 Roadmap Ingestion Pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.30",
    "pydantic>=2.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.4",
    "google-genai>=1.56.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
